Refactor logistic regression experiments: replace Hyperopt with Optuna, streamline imports, and enhance documentation
